requirements.txt
Django==3.1.*
djangorestframework==3.12.*
python-dotenv==0.15.*
psycopg2-binary==2.8.*